Move do related pis classes to dostmt package
