+++ src/main/java/org/purescript/psi/dostmt/PSDoNotationBindImpl.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.dostmt
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSDoNotationBindImpl(node: ASTNode) : PSPsiElement(node)
