+++ src/main/java/org/purescript/psi/dostmt/PSDoNotationLetImpl.kt
@@ -1,6 +1,8 @@
-package org.purescript.psi
+package org.purescript.psi.dostmt
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
+import org.purescript.psi.PSValueDeclaration
 
 class PSDoNotationLetImpl(node: ASTNode) : PSPsiElement(node) {
     val valueDeclarations get() =
