+++ src/main/java/org/purescript/psi/dostmt/PSDoNotationValueImpl.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.dostmt
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSDoNotationValueImpl(node: ASTNode) : PSPsiElement(node)
