+++ src/main/java/org/purescript/psi/expression/ExpressionIdentifierReference.kt
@@ -10,7 +10,7 @@ import com.intellij.psi.util.parents
 import org.purescript.file.ExportedValuesIndex
 import org.purescript.psi.PSLetImpl
 import org.purescript.psi.PSValueDeclaration
-import org.purescript.psi.`do`.PSDoBlock
+import org.purescript.psi.dostmt.PSDoBlock
 
 class ExpressionIdentifierReference(expressionConstructor: PSExpressionIdentifier) :
     LocalQuickFixProvider,
